Continue template decomposition: extract spotify components
